Move 2D_Soft interface in separate header
Also make the current 2D engine an "owned" unique_ptr.
